refactoring render.py
